text_summarization_gan/requirements.txt
torch
transformers
streamlit
PyPDF2
python-docx
nltk
spacy
tqdm
tiktoken
sentencepiece
protobuf
